Fix busy loop when command start is probibited
If at the start of the main loop the following conditions are met:

- tokens are available
- the plan has more ready work
- the allowed failure count is zero or
- the machine load average is over the limit

then CanRunMore() does not acquire a token and returns false. The main
loop then enters WaitForCommand() which will immediately return with
ExitTokenAvailable. That will restart the main loop and hence cause it
to busy loop until one of the conditions goes away.

Change the implementation as follows:

- move the token acquistion from CanRunMore() to the new AcquireToken()
- use the expression of the first if statement in the main loop for the
  definition of a new boolean can_run_more.
- change the condition of that first if statement so that can_run_more
  and AcquireToken() must be true to start a command.
- replace plan_.more_ready() with can_run_more in WaitForCommand().

With these changes WaitForCommand() will only monitor the token pool
when a token for executing ready work couldn't be acquired. Otherwise
the main loop falls back the original behaviour, i.e. it will only be
re-executed after a pending job has been completed.