Fix SubprocessTest.TokenAvailable on Win32
Our test code posts an I/O completion for the token but we can't control
where it ends up on the IOCP processing queue. So we need to add a loop
and call DoWork() until the token posting has been processed.

To be on the safe side we also make sure to exit the loop when the
command has completed. But if that happens the test will fail as we
expect the command to still be running.